refactor(fields): pin the curated capability set to the spec by test, not by derivation
PM adjudication on #6285 chose shape B over the runtime derivation the first
commit implemented, and this commit is that switch. Nothing about the defect or
its user-visible fix changes: `manage_sharing` still joins the curated set and
still gets its `capability.label.*` key in the ten packs and in the field
widgets' provider-less defaults map, and `labelFor` keeps the `defaultValue`
degradation.

What changes is HOW the `Mirrors PLATFORM_CAPABILITIES` claim is held.

- `CURATED_CAPABILITY_LABELS` goes back to a repo-local literal, now with
  `manage_sharing`. That keeps it readable to `check-i18n-call-site-keys.mjs`,
  whose `readVocabulary` needs a literal `new Set([…])`; the derivation made it
  `unreadable-vocabulary` and forced the family to `enumerable: false`, dropping
  the gate from 18 vocabularies / 113 exactly-checked members to 17 / 105 and
  tripping the one-way coverage ratchet in
  `scripts/__tests__/check-i18n-call-site-keys.test.ts`. Moving that ratchet is
  gate-weakening and belongs on the maintainer's floor, so it was not moved.
- `scripts/check-i18n-call-site-keys.mjs` is reverted to the base version; the
  family keeps its `kind: 'set'` vocabulary.
- The pin is renamed `specDerivation` -> `specParity` and grows the assertion
  that makes the claim mechanical: it reads the declaration through the gate's
  OWN `readVocabulary` and asserts set equality with `PLATFORM_CAPABILITIES`
  (dot -> underscore normalised) in BOTH directions, alongside the rendering
  assertions it already carried.

Measured under this shape: `check:i18n-keys` green at 18 vocabularies / 113
members — one MORE than base's 112, the extra member being `manage_sharing` —
and the ratchet test 109/109 green. The derivation remains a coherent choice and
the branch history plus the PR body keep its measurements; it simply costs a
guard, and this shape costs none.

Fixes #6285
