+++ .changeset/6285-capability-labels-derive-from-spec.md
@@ -8,18 +8,19 @@ Sharing" was the one platform capability in `sys_permission_set`'s picker that r
 English in every locale, beside seven siblings that translated — a user-visible missing
 translation, in all ten packs at once.
 
-The cause was a hand-written copy. `CURATED_CAPABILITY_LABELS` in
+The cause was an unchecked copy. `CURATED_CAPABILITY_LABELS` in
 `CapabilityMultiSelectField.tsx` listed seven capability names under a doc comment claiming
 it mirrored `@objectstack/spec/security`'s `PLATFORM_CAPABILITIES`; the spec grew an eighth
-member and the copy did not follow, so `manage_sharing` fell through to the English label
-the `sys_capability` registry serves. Nothing could catch it: the i18n gate checked that the
-seven names the copy happened to list had keys — which they did — and had no way to notice
-the member the copy never named.
+member and the list did not follow, so `manage_sharing` fell through to the English label
+the `sys_capability` registry serves. Nothing could catch it: the i18n gate reads that list
+as this key family's vocabulary and checks the members it names — all seven had keys — and
+no instrument compared the vocabulary to the array it was named after.
 
-So the set is now DERIVED from `PLATFORM_CAPABILITIES` rather than restated, applying the
-same dot-to-underscore transform the call site uses (`setup.access` → `setup_access`), and
-`capability.label.manage_sharing` is authored in all ten packs and in the field widgets'
-provider-less defaults map. A capability the spec adds next now joins the picker on the
-version bump, and a new
-`CapabilityMultiSelectField.specDerivation-6285.test.tsx` fails in CI if its label has not
-been authored — the event this card was, caught before it reaches a screen instead of after.
+`capability.label.manage_sharing` is now authored in all ten packs and in the field widgets'
+provider-less defaults map, the list carries the member, and the prose claim is replaced by
+a check: `CapabilityMultiSelectField.specParity-6285.test.tsx` imports `PLATFORM_CAPABILITIES`
+and fails on any difference in either direction, reading the declaration through the i18n
+gate's own source reader so what it pins is exactly what that gate consumes. `labelFor` also
+gains a `defaultValue`, so a capability that arrives in a future spec bump before its
+translation is authored degrades to the registry's English label rather than rendering a raw
+i18n key at the user.
